refactor: one promise-tail serializer for the daemon's file-backed stores
The six file-backed stores each carried their own copy of the same
serialize-read-modify-write primitive; json-file.ts, which already owned the
atomic write, now owns serialized(key, fn) and all six route through it. The
two repo-keyed stores' resolveRepo collapsed the same way into repo-key.ts's
resolveRepoRoot, with gitTopLevel moved beside its siblings.
