+++ packages/kobe-daemon/src/daemon/issues-store.ts
@@ -4,8 +4,8 @@ import { homedir } from "node:os"
 import { join, resolve } from "node:path"
 import { promisify } from "node:util"
 import { ROVE_STATE_DIR_BASENAME, readRoveEnv } from "../compat-env.ts"
-import { writeJsonAtomic } from "./json-file.ts"
-import { gitCommonDir, gitMainWorktree } from "./repo-key.ts"
+import { serialized, writeJsonAtomic } from "./json-file.ts"
+import { gitTopLevel, resolveRepoRoot } from "./repo-key.ts"
 
 const execFileAsync = promisify(execFile)
 
@@ -92,21 +92,13 @@ function todayStamp(): string {
   return `${d.getFullYear()}-${mm}-${dd}`
 }
 
-async function gitTopLevel(path: string): Promise<string> {
-  const { stdout } = await execFileAsync("git", ["-C", path, "rev-parse", "--show-toplevel"])
-  return stdout.trim()
-}
-
 async function resolveRepo(raw: unknown): Promise<{ repoRoot: string; repoKey: string }> {
   if (typeof raw !== "string" || raw.length === 0) throw new Error("repoRoot is required")
-  const absolute = resolve(raw)
-  const s = await stat(absolute).catch(() => null)
-  if (!s?.isDirectory()) throw new Error("repoRoot does not exist")
   try {
-    const [main, repoKey] = await Promise.all([gitMainWorktree(absolute), gitCommonDir(absolute)])
+    const { repoRoot, repoKey } = await resolveRepoRoot(raw)
     // No worktree line at all still has a readable root: fall back to the
     // toplevel rather than refuse the repo.
-    return { repoRoot: main ?? (await gitTopLevel(absolute)), repoKey }
+    return { repoRoot: repoRoot ?? (await gitTopLevel(resolve(raw))), repoKey }
   } catch (err) {
     if (isGitNotRepositoryError(err)) throw new Error("repoRoot is not a git repository")
     throw err
@@ -150,36 +142,12 @@ function response(repoRoot: string, record: RepoIssueRecord | null): RepoIssues 
   }
 }
 
-const locks = new Map<string, Promise<unknown>>()
-
-/**
- * Serialize async sections that share a resource named by `key`. The issue
- * store keeps ALL repos in one file (read/written whole), so the unit of
- * contention is the file path, NOT the repoKey — locking per-repo lets two
- * different repos' read-modify-write cycles interleave and the second
- * `writeStore` rename silently drops the first repo's mutation. Callers pass
- * `this.path` so every mutation against the file is serialized.
- */
-async function withLock<T>(key: string, fn: () => Promise<T>): Promise<T> {
-  const tail = locks.get(key) ?? Promise.resolve()
-  const run = tail.then(fn)
-  const settled = run.then(
-    () => undefined,
-    () => undefined,
-  )
-  locks.set(key, settled)
-  void settled.then(() => {
-    if (locks.get(key) === settled) locks.delete(key)
-  })
-  return run
-}
-
 export class IssuesStore {
   constructor(private readonly path = defaultIssuesStorePath()) {}
 
   async list(repo: unknown): Promise<RepoIssues> {
     const { repoRoot, repoKey } = await resolveRepo(repo)
-    return withLock(this.path, async () => {
+    return serialized(this.path, async () => {
       const store = await readStore(this.path)
       const record = store.repos[repoKey] ?? null
       if (record && record.repoRoot !== repoRoot) {
@@ -203,7 +171,7 @@ export class IssuesStore {
   async mirrorTaskDone(repo: unknown, taskId: string): Promise<RepoIssues | null> {
     const { repoRoot, repoKey } = await resolveRepo(repo)
     if (!taskId) return null
-    return withLock(this.path, async () => {
+    return serialized(this.path, async () => {
       const store = await readStore(this.path)
       const record = store.repos[repoKey]
       if (!record) return null
@@ -228,7 +196,7 @@ export class IssuesStore {
   async unlinkTask(repo: unknown, taskId: string): Promise<RepoIssues | null> {
     const { repoRoot, repoKey } = await resolveRepo(repo)
     if (!taskId) return null
-    return withLock(this.path, async () => {
+    return serialized(this.path, async () => {
       const store = await readStore(this.path)
       const record = store.repos[repoKey]
       if (!record) return null
@@ -246,7 +214,7 @@ export class IssuesStore {
     if (!op || typeof op !== "object" || Array.isArray(op) || typeof (op as { type?: unknown }).type !== "string") {
       throw new Error("missing op")
     }
-    return withLock(this.path, async () => {
+    return serialized(this.path, async () => {
       const store = await readStore(this.path)
       let record = store.repos[repoKey]
       if (!record) {
